editing the file 2-append_text_to_file.c

+++ file_io/2-append_text_to_file.c
@@ -11,7 +11,7 @@ int append_text_to_file(const char *filename, char *text_content)
 {
 	int fileDescriptor;
 	int len;
-	ssize_t bytesWritten;
+	int bytesWritten;
 
 	if (filename == NULL)
 		return (-1);
